Feat: Create cart controller & service & dto

+++ src/main/java/com/myomi/board/repository/BoardRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.myomi.board.entity.Board;
 
-public interface BoardRepository extends CrudRepository<Board, Integer> {
+public interface BoardRepository extends CrudRepository<Board, Long> {
 
 }
